requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==3.2.16
django-chartjs==2.3.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Faker==15.3.1
geographiclib==1.52
geopy==2.2.0
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.5
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
